+++ src/main/java/com/atlassian/db/replica/internal/aurora/AuroraReplicasDiscoverer.java
@@ -14,9 +14,11 @@ import static java.util.stream.Collectors.toList;
  */
 public final class AuroraReplicasDiscoverer {
     private final AuroraJdbcUrl readerUrl;
+    private final boolean ignoreInactiveReplicas;
 
-    public AuroraReplicasDiscoverer(AuroraJdbcUrl readerUrl) {
+    public AuroraReplicasDiscoverer(AuroraJdbcUrl readerUrl, boolean ignoreInactiveReplicas) {
         this.readerUrl = readerUrl;
+        this.ignoreInactiveReplicas = ignoreInactiveReplicas;
     }
 
     /**
@@ -38,10 +40,11 @@ public final class AuroraReplicasDiscoverer {
 
     private List<String> fetchReplicasServerIds(Connection connection) throws SQLException {
         List<String> ids = new LinkedList<>();
+        final String sql = ignoreInactiveReplicas ?
+            "SELECT server_id FROM aurora_replica_status() WHERE session_id != 'MASTER_SESSION_ID' and last_update_timestamp > NOW() - INTERVAL '5 minutes'" :
+            "SELECT server_id FROM aurora_replica_status() WHERE session_id != 'MASTER_SESSION_ID'";
         try (ResultSet rs =
-                 connection.prepareStatement(
-                     "SELECT server_id FROM aurora_replica_status() WHERE session_id != 'MASTER_SESSION_ID'")
-                     .executeQuery()) {
+                 connection.prepareStatement(sql).executeQuery()) {
             while (rs.next()) {
                 ids.add(rs.getString("server_id"));
             }