+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-2.6.2...master
 
+### Fixed
+- Ignore replica nodes when the topology state for the node was not updated for 5 minutes.
+
 ## [2.6.2] - 2022-04-20
 [2.6.2]: https://github.com/atlassian-labs/db-replica/compare/release-2.6.0...release-2.6.2